run bad token filtering and liquidity fetching concurrently (#3407)
# Description
A test where we disabled our in cluster nodes showed that we still lose
a lot of time when we don't have high performance nodes. One thing that
became apparent was that we should filter bad tokens and fetch liquidity
concurrently.

EDIT: after further investigation the time savings for doing this
concurrently are probably not that significant after all but it wouldn't
hurt either so I'll leave the PR open.

# Changes
do the 2 tasks concurrently

## How to test
e2e should still pass
performance test on shadow would be nice