[EASY] Current block staleness metric (#3426)
# Description
While investigating one of the issues, it turned out that we don't have
good observability if the current block watcher is having difficulties
fetching new blocks. The current `block_stream_update_delta` records
only the delta between blocks, but sometimes the watcher is unable to
receive a new block for many minutes, while the RPC metric is fine.
While this PR doesn't fix the root cause(which is currently unknown), a
Grafana panel and an alert should help to speed up investigations of
other issues that can happen because of that.

# Changes
A new metric that records the latest observed block number on each
current block watcher iteration, so it should be visible for how long we
see the same block. The metric in Prometheus will have a service prefix,
like `api_`, `driver_`, etc.

## How to test
Grafana.