[EASY] Add metric for multiple winners (#3428)
# Description
To better track the behavior and impact of combinatorial auctions, we
want a new metric that allows us to visualize the amount of auction
winners.

We want to plot a dashboard panel to group auctions by number of winners
and compare the amount of auctions in each group.

# Changes
A new `Histogram` metric `gp_v2_autopilot_runloop_auction_winners` that
records how many winners did an auction have

## How to test
Tested on staging 

<!--
## Related Issues

Fixes #
-->