Abort driver settlement task when autopilot drops the `/settle` request (#3427)
# Description
An issue was recently observed where the settlement queue remained full
for a long period of time. That was caused by a stuck current block
stream, which blocked the settlement task in the driver. While the
autopilot had already aborted the `/settle` request, due to the nature
of the async settlement operation, it continued to execute until the
current block stream was recovered. During that time, the settlement
queue could no longer accept any additional requests.

One possible(and very trivial) option to address it is to set a next
block awaiting timeout, but that covers only this specific async
operation, where many more potentially can block the queue. This PR
introduces a mechanism to detect whether the `/settle` request is
dropped, which cancels the settlement task and also stops the internal
loop that awaits the settlement tx.

# Changes
In the function that receives a new settlement execution request from
the settlement queue, a `response_sender` is already provided, which is
used as a channel to return results to the `/settle` endpoint caller.
When the client drops the request, the `response_receiver` is also
dropped. This can be detected by awaiting the `response_sender.closed()`
execution. This exact event is used as a cancellation signal for the
settlement task. The task will be finished and removed from the
settlement queue.

For a reference, a place where the response sender and receiver are
initialized. So, dropping `response_receiver` will trigger the
implemented signal:

https://github.com/cowprotocol/services/blob/79b129f41891393743d1789a59e98da7f492e6f1/crates/driver/src/domain/competition/mod.rs#L341-L368

## How to test
That seems to be very tricky, so probably ensuring it works properly in
staging should be enough.