Fix sorting of ranked solutions (#3429)
# Description
The `Arbitrator` trait expects the solutions to be sorted in a
particular way: winners before losers, high scores before low scores.
This is currently not enforced. A wrong ordering will result down the
line with solutions being stored in the wrong order in the solver
competitions table. This results in the possibly wrong ordering in the
solver competitions and order status API endpoints.

Effectively this now also ensures that the order status endpoint will do
the right thing when we switch to combinatorial auctions.

# Changes
Implement correct sorting in the combinatorial auctions implementation.
Assert ordering in unit tests.
Also adjusted unit tests to enforce winners to be specified in that same
order as well. This was not needed but seems nicer for consistency. No
such ordering was enforced for `expected_fair_solutions` since the trait
does not expect any particular ordering for that.
Lastly I switched from unstable to stable sorting algorithms just to
make things more reproducible.

## How to test
The test changes in the PR are very circular (same code that sorts is
also used to check order) but I felt it was not worth it to make the
asserting code more verbose to make the correctness explicit.
Instead I wrote a small rust [playground
test](https://play.rust-lang.org/?version=stable&mode=debug&edition=2024&gist=67cbf46bcc14232f031e8acafa35a78a)
to show that the sorting is indeed correct.