Detect under funded flashloans (#3432)
# Description
The driver aims to abstract flashloan handling for orders completely
away from the solver (if they don't want to handle that themselves).
That means the driver should also offer a reasonable amount of
protection.

# Changes
Added a sanity check in the driver to verify that the funds paid by
flashloan orders indeed cover the repayment cost of the flashloans.
It's admittedly a bit ugly to only detect this case when we encode the
solution but that's currently the only place where we have all the
necessary data.

## How to test
Initially adjusted the flashloan e2e test to trigger the issue where the
driver allows executing flashloans for orders which don't cover the
repayment cost.
After the fix was implemented the driver refused to submit these
problematic solutions.