Execute hooks of ethflow orders (#3433)
Fixes: https://github.com/cowprotocol/services/issues/3388

# Description
Currently ethflow orders don't have their hooks executed. The reason is
that ethflow orders don't get added to the order book via a REST API
call which would execute the usual code paths. Instead the autopilot
indexes events emitted from certain contracts and adds it to the
orderbook that way. This executes a bunch of different code paths that
actually didn't check the hooks at all.

# Changes
- moved `--hooks-contract-address` into shared args because the
`autopilot` now needs it too
- during ethflow event indexing we now fetch the associated appdata,
parse it, and insert the trampolined hook interactions if there are any.
This code is a bit awkward due to the existing abstractions. Ethflow
orders are a special variant of the more generic "onchain" orders. The
ethflow specific handler already inserts the necessary unwrap
pre-interaction. Because the general appdata handling needs to happen in
the general handler and not the ethflow handler we first need to check
how many interactions were already inserted for a given order so we
don't overwrite any interactions by accident. Because the majority of
times there is only 1 ethflow order per tx and to keep the code simple I
decided to fetch the appdata sequentially instead of in parallel.
- because ethflow orders can be reorged (as they originate from onchain
transactions) we need to be able to overwrite interactions in the DB. To
do that I needed to adjust the primary key of the `interactions` table
to include the `execution` flag. The readme was updated accordingly.

## How to test
I updated an existing ethflow e2e test to pass 1 pre and post hook in
the appdata. The hooks set an allowance for the trader address. Because
we use approval interactions we can easily see in which context the user
passed interaction got executed.
If we didn't sandbox the interaction correctly the approval would be set
by the settlement contract. But in our test we can see that the approval
was given by the trampoline contract. This shows that we didn't
introduce a security risk by accident with this fix.