fix: mismatch between onchain settlements and database solutions on non partially-fillable orders (#3440)
# Description
On #3353 we implemented matching logic to correlate onchain settlement
trades with the solutions stored in database. This was done to pave the
way for combinatorial auctions.

Recently we detected an edge case that happens when the order is NOT
partially fillable. The Settlement smart contract
[[1]](https://github.com/cowprotocol/contracts/blob/main/src/contracts/GPv2Settlement.sol#L385-L386)
[[2]](https://github.com/cowprotocol/contracts/blob/main/src/contracts/GPv2Settlement.sol#L407-L408)
does override the `executed` amount with the actual sell/buy amounts
(including fees). This in turn makes the current implementation unable
to match the onchain trade with the database entry for the corresponding
solution, resulting in a `InconsistentData::SolutionNotFound`.

This PR updates the onchain-db matching logic so we take into account
the Settlement smart contract behavior.

Also tagged as `hotfix` to be applied immediately to production.

# Changes
- [ ] Updated the logic behind the `find_winning_solution_uid` used in
the settlement observer, to treat the special case of a non
partially-fillable trade.

## How to test
Added a new unit test `settlement_with_fill_or_kill` that reproduces the
issue.

<!--
## Related Issues

Fixes #
-->

---------

Co-authored-by: Martin Magnus <martin.beckmann@protonmail.com>