chore(OpenAPI): SolverCompetition transactionHash should be an array (#3442)
# Description
The OpenAPI spec is not up to date with the data returned in the
corresponding endpoint, as the endpoint returns multiple transaction
hashes to support multiple winners (ref #3049) but the OpenAPI still
defines it as a single value.

This PR updates the field to be an array and updated the name to match
the field.

<!--
## Related Issues

Fixes #
-->

Co-authored-by: Martin Magnus <martin.beckmann@protonmail.com>