Filter out problematic orders instead of discarding the auction (#3449)
# Description
Under some conditions the driver rejects auctions with malformed orders
altogether. This led to all solvers connected to the reference driver
not participating for 30 minutes while such an order was part of the
auction.
The
[order](https://explorer.cow.fi/base/orders/0xfcbd64398f0a03405b4c37f356a36bc0acc538fa144f67a9471bafd905712d62ba3cb449bd2b4adddbc894d8697f5170800eadecffffffff)
in question has a buy amount of 0 which caused the `InvalidAmounts`
error.

# Changes
Instead of discarding the entire auction we can simply filter out the
problematic orders.