New solver competition endpoint (#3453)
The `/solver_competition` endpoint is used by many solver teams to
monitor how well their solver is doing compared to others. With
combinatorial auctions there are a few pieces of information that would
be very useful to have in there:
* reference scores as part of solutions of a winning solver
* tx_hash of executed solutions

Currently the `/solver_competition` endpoint mostly loads a big JSON
stored in the DB and adds very few pieces of information on top. Since
this is HORRIBLE to work with we prepared multiple tables that store all
that information in a reasonable and well defined layout (instead of 1
huge JSON). This PR finally start using these new tables by
reimplementing the `/solver_competition` endpoint with them.

The ONLY relevant thing the new endpoints currently don't have is a
readable solver name associated with each solution. The only place where
we strictly need these solver names at the moment is for the order
progress bar. Once we either add these names to the solutions or have
the frontend find some other way to get readable names we can drop the
old endpoint (or replace the implementation).

This PR introduces a new version of the 3 existing `/solver_competition`
endpoints (latest, by id, by transaction hash).
Because they use completely new tables for this I introduced `_v2`
versions of many files dedicated to the new logic. I also migrated some
existing logic that uses the new tables into the new `_v2` database
file.

Updated e2e tests to use the new endpoints instead of the old ones.
Added assertions for the new fields in the endpoint.

---------

Co-authored-by: ilya <ilya@cow.fi>