Enable trace collection (#3448)
# Description

Initial setup to export tracing information (spans). Currently requests
IDs are being propagated via X-Request-ID header semi-manually and then
added to logs. It is a working solution, but with Grafana Tempo we can
collect traces in open telemetry format and have a waterfall graph in
Grafana showing execution flow. When we propagate tracing context via
http calls we can use a single trace ID to show the entire lifetime of a
flow. This is a great profiling tool, because one can see where we're
spending the most time from looking at the graph.

This MR only sets up the config structs and add a new tracing layer that
collects and batch exports traces to Tempo.

# Changes
- [x] Added a config struct for observability
- [x] Set up a tracing export layer for Grafana Tempo
- [ ] Make tracing distributed (propagate context)
- [ ] Add trace IDs to logs for easy correlation

## How to test
<!--- Include details of how to test your changes, including any
pre-requisites. If no unit tests are included, please explain why and
how to test manually
1.
2.
3.
-->

<!--
## Related Issues

Fixes #
-->