Add distributed tracing support (#3463)
Add context propagation over http for servers (axum and warp) and
reqwest calls.


# Description
o be able to view full execution flow over different services aggregated
nicely in a graph in Grafana we need to propagate tracing context in
HTTP headers to downstream services so tempo knows these execution flows
belong together.

<img width="1445" alt="Screenshot 2025-07-02 at 11 09 35"
src="https://github.com/user-attachments/assets/2405160a-b8fb-4651-8910-65fc3fa42df6"
/>



# Changes

- [x] Added a formatting layer that grab a trace ID a appends it to all
JSON logs and human-readable logs, so one can grab that ID and search
for the execution flow in Grafana. This is done by introducing custom
formatters that format log events as one normally would, but attach a
trace ID to the output.

Example of human-readable logs:
```
[trace_id=4bf92f3577b34da6a3ce929d0e0e4736] 2025-07-04T13:35:17.741Z  INFO  http_request{method=GET path=/api/v1/version request_id=123}: warp::filters::trace: processing request
```


Example json:
```json
{
  "timestamp": "2025-07-04T12:58:56.138095625+00:00",
  "level": "INFO",
  "fields": {
    "message": "finished processing with success",
    "status": 200
  },
  "target": "warp::filters::trace",
  "trace_id": "4bf92f3577b34da6a3ce929d0e0e4736"
}
```

- [x] Added context extraction/injection helper functions for axum, warp
& reqwest
- [x] Added a new config struct for distributed tracing - tracing is
switched on when a collector endpoint is provided
- [x] Tracing level is now configurable 


## Custom formatters

This MR introduces two custom tracing event (aka "log") formatters:
TraceIdJsonFormat and TraceIdFmt. A formatter is a struct that takes in
an event and its context and produces serializes it for output. The
purpose of these new formatters is to add tracing IDs to all logs. The
human-readable (non-JSON) formatter is a wrapper around the standard
event formatter from the tracing lib. It first writes the trace ID to
the output and offloads the rest to the standard formatter; we use this
formatter in prod. The JSON formatter, which we don't really use
anywhere atm, produces a JSON output which is supposed to be more easily
parsable and thus easier for machine processing. One can't really "just
append" some text and get a valid json object, so this formatter does
all the work of serializing/encoding events. It behaves like the [JSON
formatter shipped with the tracing
library](https://github.com/tokio-rs/tracing/blob/main/tracing-subscriber/src/fmt/format/json.rs#L228)
with `display_span_list` and `current_span` flags off, but adds a
tracing ID to the output.

---------

Co-authored-by: Martin Magnus <martin.beckmann@protonmail.com>
Co-authored-by: ilya <ilya@cow.fi>