Fix logs not being colored (#3491)
# Description

We're wrapping the default formatter and the ansi flag was not being
propagated from Layer to Formatter which then sets the flag on Writer.
This led to log output not using colors. By setting the flag on the
Format struct the event formatting code will set it on the writer as
well:
https://github.com/tokio-rs/tracing/blob/main/tracing-subscriber/src/fmt/format/mod.rs#L922.