Allow placing orders when their flashloan covers them (#3494)
# Description
There are plenty of use cases where it makes sense to allow created
orders which are currently not covered (owner doesn't have the funds
right now). Usually we rely on a simulation based approach to figure out
whether an order is actually funded. Adjusting this logic for flashloans
is a lot of work and probably not that useful to begin with.

# Changes
Instead we simply inspect the flashloan hint and make sure that it would
flashloan the required funds to the user.

## How to test
Added a unit test

kudos to @mrnaveira