Implement instrumentation layer with alloy (#3472)
# Description
To aid a piecewise migration towards alloy we need to start by making
the alloy `Provider` equivalent to our current `Web3` provider.
This PR implements the instrumentation aspect of our current transport.
As a reminder this is what we want to support:
* providers can be "labeled" - every request coming through this
provider should get this label
* you can take a labeled provider, attach another label to it, and get a
new provider with both labels on it
* we collect metrics for each RPC request based on the provider it came
from

# Changes
- bump `itertools` to `0.14` to avoid a version conflict with `alloy`
    - replace deprecated `group_by` call with equivalent `chunk_by`
- add `alloy` as a workspace dependency pinned to this upstreamed
[commit](https://github.com/alloy-rs/alloy/pull/2672) because we need it
for this PR (once this gets released we can pin the released version)
- copy `Metrics` code from our current `web3` transport
- that way we can collect the identical metrics under a different name
to compare the new alloy and the current transport

Implement `LabelingLayer` and `Provider`
This layer simply adds it's own label to the `ProviderLabel` of every
request passing through. This is needed to support the use where we want
to build a hierarchy of providers to get fine grained metrics for
individual components. For example the "main" provider can be wrapped in
a "liquidity fetching" provider. Now all requests coming through the
wrapping provider will get both labels. Note that the request first
passes through the "liquidity fetching" provider and then through the
"main" provider so their labels get added in the reverse order.

Implement `ProviderLabel`
This is effectively just a `String` that is aware that labels get added
in the reverse order and therefor it prepends instead of appends. The
current implementation is not optimal but without benchmarking it's not
clear how big of an impact this will actually have. I decided to keep
the code simple for now.
At least the label overallocates capacity in the string to avoid
reallocations when adding more labels.

Implement `InstrumentationLayer` and `Provider`
For every incoming request it tries to read the associated
`ProviderLabel` (if present) and collects the same metrics as our
current web3 transport layer.

## How to test
I mostly ran some manual tests. Before starting to actually use the new
alloy providers I'll add a separate PR for unit tests.