Fix `solver_competition::load_by_id` SQL query (#3485)
# Description
The following log was observed when the settlement's `solution_uid` is
not yet updated:
```
2025-07-12T09:06:24.160Z ERROR request{id="3b3dca946e4113162fefcd6c1563e6b1"}: orderbook::api::get_solver_competition_v2: load solver competition err=solver_competition::load_by_id

Caused by:
    0: error occurred while decoding column "solution_uid": unexpected null; try decoding as an `Option`
    1: unexpected null; try decoding as an `Option`
```
This PR fixes that by handling this column as optional. Also fixes
`solver_competition_v2::load_latest` and
`solver_competition_v2::load_by_tx_hash` since they both use
`load_by_id` internally.

# Changes

- [ ] Fix the SQL query. The `solution_uid` is populated separately, so
it still can be NULL. The query checks for that since solution's tx hash
is not returned in the API response until the uid is updated in the DB.

## How to test
New tests.