Fix `stderr_threshold` setting (#3498)
# Description
Follow up item of a post mortem.
During the post mortem it was discussed that we did not get any slack
alerts despite emitting logs with the `error` log level. It was stated
that it was a deliberate decision to stop emitting these errors as slack
alerts due to noisy. However, my investigation suggests that this was
not deliberate and the functionality simply broke in #2868.

The infra repo still configures `stderr_threshold` as `error` and the
`stderr_threshold` config was never optional so some number of logs was
always expected to go to stderr.

# Changes
- fixed the bug that caused us to stop writing to stderr for logs that
had the correct level
- made `stderr_threshold` optional and disabled by default
- converted `LevelFilter` config member arguments to `Level` as that can
better express optionality of arguments. You can't use `LevelFilter` to
express that you want to always want logging to happen because you can
populate it with `LevelFilter::Off` which disables logging, that's not
possible with `Level` because that does not have the `OFF` variant. To
express optionality with `Level` you have to use `Option<Level>` which
communicates intent very clearly.

## How to test
Ran e2e tests with test logs with different levels and
`stderr_threshold` settings, piped `stderr` to a file and verified that
it only contains logs that were supposed to show up.