Implement auto batching transport layer (#3473)
# Description
Implements our custom auto batching transport layer in alloy. Our
current batching layer also manages and propagates the `x-request-id` to
trace individual calls back to the user request that we initiated it
for. This introduced a lot of complexity and is not really related to
the raw batching. This will be tackled in a separate PR but IMO wouldn't
even block the start of the migration.

# Changes
Reimplements the core logic of the current [buffering
implementation](https://github.com/cowprotocol/services/blob/main/crates/ethrpc/src/buffered.rs).
The idea is that the batching provider simply forwards individual calls
into a queue that a background task reads from. That background task
then aggregates individual calls into batches and then forwards them to
the inner provider that is tasked with eventually sending the HTTP
request.
Once the results are received the background task will send them back to
the original provider through a oneshot channel.

The original transport also handles manually batched requests. Since
manually batching is usually harder to read and not needed with this
layer I decided to not handle that case and add a custom clippy lint
complaining when manual batching is used.

## How to test
So far only a manual test but I'll add a unit test PR before actually
starting to use this in our code.