Use custom storage layout for quote verification helper contracts (#3500)
# Description
There was a peculiar bug where quote verifications for solvers which
were smart contracts instead of EOAs would fail. The
[simulation](https://www.tdly.co/shared/simulation/8083f0bb-235a-4a90-95a9-c74ee6700d27)
ends with thousands of `SLOAD`s until we reach the gas limit.
The reason for this is actually super interesting. For our quote
verification we need to be able to make the solver and user address to
things which they would normally have to sign off of. To make that
happen we deploy helper contracts at the solver and user address with
functions that implement the desired behavior that we can simply call.
We "deploy" these helper contracts by putting the helper contract
bytecode at the respective addresses by overwriting the accounts' code
using state overrides. When the solver account is an EOA that worked
because EOAs don't have any storage associated with them (also not
completely true anymore since EIP 7702 but that's beside the point). The
particular smart contract solver that caused issues actually had the
storage slots where our helper contract stores its information populated
with some value.
The consequences are the following:
`_simulationOverhead` would contain the actual simulation overhead +
whatever number the contract had stored at the associated storage slot.
And `_queriedBalances` (an array) would appear to have way more entries
because the actual contract had some value stored in the storage slot
that stores the length of the array. When the logic tries to copy the
`_queriedBalances` into the return value it tried to load thousands of
uninitialized storage slots.

# Changes
I tested this theory be resimulating the failing quote verification with
the array length storage slot explicitly set to zero and it
[worked](https://www.tdly.co/shared/simulation/667d4c53-3108-4825-895c-156f50fc7210).

The change is very simple: we use a new solidity feature to make our
contracts use a custom storage layout instead of the default. This
should prevent any accidental collisions while preserving the storage of
the real contract in case it's needed during the settlement for some
reason.
- bumped `solc` version to `0.8.30` in the make file
- used custom storage layouts for `Solver.sol` and `Trader.sol`
- removed custom storage slot logic in `Trader.sol` as it's not needed
anymore
- bumping the compiler version also generated a new (hopefully better
optimized) bytecode for `Spardose.sol`
- the bumped compiler version also targets a different hardfork by
default, to stay compatible with e2e tests forked off of old blocks we
explicitly target the `shanghai` fork when compiling the contracts

thanks @fedgiac for the custom storage layout feature recommendation

## How to test
Indirectly tested via the adjusted tenderly simulation. This can
actually be tested in an e2e test but given that this is needed quickly
and the change has 0 risk I'll create a separate PR for the test later.