Add deployment addresses for flashloan contracts (#3495)
# Description
So far we freshly deployed the 2 helper contracts related to flashloans
even in forked tests because they haven't been audited and deployed when
the test code was written.
Now they are deployed on all our networks so we can update the test
setup code where we "look up" contracts that were deployed.

Unfortunately we have a couple of forked e2e tests which rely on state
before these contracts were deployed so for now I made the contracts
optional to handle that case too.

# Changes
- added deployment addresses for the deployed contracts. This did not
happen for the maker contracts as that was not widely deployed. Luckily
it's currently irrelevant for our testing.
- adjusted our e2e setup code to deal with these optional contracts

## How to test
I bumped the block number for the flashloan test to a block which
already had the helper contracts deployed.
That required a small adjustment of one of the asserts due to the
different state of the chain on the new block.