Fix memory leak on solver background task timeout (#3499)
# Description
Currently we do not clean up timed out background tokio tasks on the
solvers. So when there is a high number of timeouts the memory usage
keeps increasing and is never released, which can result in Out of
Memory errors.

This PR fixes it by explicitly aborting the tokio task in case of a
timeout.

## How to test
This change was tested on the baseline solver on mainnet shadow, with a
configuration (uni v3) enabled that previously increased the usage until
the pod was OOMKilled, but with this change the memory usage is stable.