Fix quote verification for tokens which don't allow 0 approvals (#3504)
# Description
There were some tokens for which we failed to verify quotes. The reason
turned out to be that these tokens revert if you set the approval to 0.
However, this is needed for tokens like USDT as that is used to protect
against an exploit.

To support both of these tokens at the same time we can simply catch any
reverts coming from the approve calls.

# Changes
- made the vendored `safeApprove` function `external` as that was
required to be able to use `try catch` around it
- wrapped both `safeApprove` calls in try catch and ignored any reverts

## How to test
I took a quote verification
[simulation](https://dashboard.tenderly.co/cow-protocol/production/simulator/1d8ea577-2ea3-4bbf-ac43-782f7ed62661)
that failed, and sparked this investigation and replaced the trader
bytecode with the new trader code from this PR.
The adjusted
[simulation](https://dashboard.tenderly.co/cow-protocol/production/simulator/1a3a22cf-fdaa-4113-affb-7b7c4bd3229a)
is able to get past the trader setup logic that is specific to the quote
verification.

However, it still failed when executing the interaction where the
settlement contract sets an allowance for some liquidity pool. That is
an interaction that is not related to the quote verification and gets
encoded
[here](https://github.com/cowprotocol/services/blob/main/crates/driver/src/domain/quote.rs#L311).
That means the solution returned by the reference driver is currently
also not compatible with these tokens. Fixing that incompatibility is
probably more involved. AFAIU it's not possible to just have another
helper contract that handles reverts. I believe the driver instead has
to figure out the actual behavior of the token and encode the
interactions accordingly. So for USDT it would mean to encode 2
approvals (0, max) while for the new problematic token it should only be
the max approval.

I think this PR still makes sense since theoretically colocated solvers
which don't use the reference drivers would at least be able to provide
quotes that can be verified (if they have this special handling
implemented in their custom driver).