[EASY] Do not solve empty auctions in driver (#3507)
# Description
While testing locally, I spent a lot of time understanding why the
solver receives weird requests during the solve competition, but it
works fine for quoting. It turned out that if the prioritization task in
driver filters out all the orders, the competition continues and in the
logs we just see NoSolutions, which is confusing.

# Changes
This PR addresses it by returning fast a proper response for this case.

## How to test
Local runs.