Fix memory leak on autopilot when solving (#3505)
# Description
Following up on https://github.com/cowprotocol/services/pull/3499, there
is another part of the code where do not clean up timed out background
tokio tasks. So when there is a high number of timeouts, potentially
this could end up in a OOM situation.

This PR fixes it by explicitly aborting the tokio task in case of a
timeout. Note that this is purely preventive, we don't have record of
this happened yet on any environment

## How to test
No way of testing other that all the existing CI tests passing