Simulate balances using state overrides (#3506)
# Description
Fixes: https://github.com/cowprotocol/services/issues/3431

The hope is that this is compatible with lens as contract deployment
seems to work differently on `zksync` chains.

# Changes
Implements logic to compute user balances using state overrides instead
of constructor magic. Explained in detail in the linked issue.

This is the first real use of `alloy` as that is far superior to
`ethcontract-rs` when it comes to decoding raw bytes into reasonable
types.

## How to test
I ran the tests on a temporary commit that runs both implementations and
asserts that the outputs are identical. This was the assertion:
```rust
    async fn simulate(&self, query: &Query, amount: Option<U256>) -> Result<Simulation> {
        let (v1, v2) = tokio::join!(
            self.simulate_v1(query, amount),
            self.simulate_v2(query, amount)
        );
        if v1.is_err() != v2.is_err() {
            tracing::error!(?v1, ?v2);
            panic!("ABORT");
        }
        if let (Ok(v1), Ok(v2)) = (&v1, &v2) {
            assert_eq!(v1, v2);
        }
        v1
    }
```