Revert "Simulate balances using state overrides" (#3518)
Reverts cowprotocol/services#3506

While this code appears to be saner it appears to have caused a
considerable performance regression. Since the regression was very
surprising to me I opened a
[discussion](https://github.com/paradigmxyz/reth/discussions/17593#discussioncomment-13872743)
on the reth repo about it. You can find more details there.

Note that this or a different solution is still needed to make the code
work on lens. So tagging @squadgazzz here explicitly to not get
surprised when lens stops working after merging main into the `lens`
branch.