[TRIVIAL] Make `stderr` log level configurable in driver (#3513)
# Description
https://github.com/cowprotocol/services/pull/3498 overlooked that the
driver does not use the shared logging args struct. Instead it has it's
own arguments. Since the driver didn't have an argument for
`stderr_threshold` it always set the level to `error` (meaning all
`error` logs show up as alerts on slack).

# Changes
made the `stderr_threshold` configurable in the driver by adding
`--std--threshold` CLI argument. The feature is disable by default which
is in line with the other binaries.