Don't fetch future block on autopilot restarts (#3521)
# Description
We are currently seeing sporadic panics (around 50% of the time) when
the mainnet autopilot restarts in production. I believe this is due to
us trying to fetch "last seen block + 1" to initialize the eth flow
refund (and settlement) event indexing flow. In particular what happens
is:

1. The old autopilot instance sees a new block _t_ with a relevant event
(~every other block for mainnet production)
2. Autopilot restarts and now tries to fetch block _t+1_ from the RPC,
which doesn't yet exist (12s block production time) and thus panics
3. After the next restart it finds the block t+1 and starts correctly

This is causing sporadic micro outages (~30s) on mainnet production

https://github.com/cowprotocol/services/pull/3480 attempted to fix the
same issue by introducing a retry mechanism. However, given the chosen
parameters (3 retries with 500ms delay) this likely has only little
effect on mainnet, where we would have to retry for 10+ seconds in order
to see the next block

# Changes

- [x] remove retry logic to keep code simple
- [x] remove the +1 from the block fetching logic (I checked that
potential reorgs are handled in the core event updater logic)
- [x] bubble up error message from `block_number_to_block_number_hash`
to better debug what the node is responding in the error case

## How to test
This is hard to test since on staging we are barely having settlements
(and thus last seen block is usually late in the past). Any suggestion
on how to get more confidence in the change is appreciated.