chore(baseline): rename node_url config to uni_v3_node_url  (#3517)
# Description
On https://github.com/cowprotocol/services/pull/3468 we added support
for UniV3 LPs in the Baseline solver. It introduced a new `node_url`
configuration parameter in the baseline solver, that specifies the RPC
URL to use for UniV3 and is also used to enable/disable the feature.

To improve readability of the baseline solver configuration (e.g. to
quickly identify/react to any alert or emergency), this PR renames the
configuration parameter to `uni_v3_node_url`, to clearly identify that
it manages the UniV3 feature.