Use deployed helper SC to estimate tradable balances (#3550)
# Description

ZKSync currently has the following limitation:`delegatecall` Is Not
Supported Between EVM and EraVM
Contracts[[link](https://docs.zksync.io/zksync-protocol/evm-interpreter/era-evm-interactions#delegatecall-is-not-supported-between-evm-and-eravm-contracts)]

This PR is the outcome of the #3545's discussions, which was another
attempt to mitigate the above, with the following conceptual change.
Instead of using different workarounds depending on the chain type, come
up with a solution that works everywhere in order to reduce the code
complexity.

So far, 3 solution types were discovered:
1. Use the `storage_accessible` module, which doesn't work on ZKSync due
to the absence of EVM<->EraVM `delegatecall` support.
2. Use state overrides, which was initially proposed in #3506, and even
though it was the most concise solution, it was reverted in #3518 due to
performance issues.
3. Actually deploying the helper SCs and using them apart from the
`storage_accessible` module.

The last option was selected, since it shouldn't introduce any
performance issues and was already tested on Lens.

# Changes

- [ ] Deploy the Balances smart contract instead of generating it on the
fly.
- [ ] Adapt driver and e2e tests.

## How to test
Existing tests. I was a bit surprised that the forked node e2e tests
work as is without issues since the Balances SC is deployed after the
currently set forked blocks in some tests.